feat: isolate `hardforks` into a crate without tempo deps
