+++ .github/workflows/semver-check.yml
@@ -6,6 +6,7 @@ on:
   pull_request:
     branches: [main]
     paths:
+      - crates/hardfork/**
       - crates/contracts/**
       - crates/primitives/**
       - crates/chainspec/**
