+++ crates/chainspec/Cargo.toml
@@ -17,6 +17,7 @@ publish = true
 workspace = true
 
 [dependencies]
+tempo-hardfork.workspace = true
 tempo-primitives.workspace = true
 
 reth-cli = { workspace = true, optional = true }
@@ -29,12 +30,9 @@ alloy-genesis.workspace = true
 once_cell.workspace = true
 alloy-primitives.workspace = true
 alloy-eips.workspace = true
-alloy-hardforks.workspace = true
-
 eyre = { workspace = true, optional = true }
 serde.workspace = true
 serde_json.workspace = true
-paste.workspace = true
 commonware-codec = { workspace = true, optional = true }
 commonware-cryptography = { workspace = true, optional = true }
 
@@ -43,6 +41,7 @@ commonware-cryptography = { workspace = true, optional = true }
 [features]
 default = ["std", "serde", "reth", "cli"]
 std = [
+    "tempo-hardfork/std",
     "alloy-eips/std",
     "alloy-evm?/std",
     "alloy-genesis/std",
@@ -55,12 +54,12 @@ std = [
     "tempo-primitives/std",
 ]
 serde = [
+    "tempo-hardfork/serde",
     "alloy-eips/serde",
-    "alloy-hardforks/serde",
     "alloy-primitives/serde",
     "tempo-primitives/serde"
 ]
-evm = ["dep:alloy-evm", "tempo-primitives/evm"]
+evm = ["dep:alloy-evm", "tempo-hardfork/evm", "tempo-primitives/evm"]
 reth = [
     "std",
     "evm",
