+++ crates/contracts/Cargo.toml
@@ -21,6 +21,7 @@ alloy-contract = { workspace = true, optional = true }
 alloy-primitives.workspace = true
 alloy-sol-types = { workspace = true, features = ["json"] }
 serde = { workspace = true, optional = true }
+tempo-hardfork.workspace = true
 
 [dev-dependencies]
 alloy-primitives = { workspace = true, features = ["rand"] }
