+++ crates/contracts/src/lib.rs
@@ -9,6 +9,8 @@
 #[cfg(test)]
 extern crate alloc;
 
+pub use tempo_hardfork::TempoHardfork;
+
 use alloy_primitives::{Address, B256, address, b256};
 
 /// Default address for the Multicall3 contract on most chains. See: <https://github.com/mds1/multicall>
